Trigger a new push build
The IntegrationTestScenario for the last push did not run. As a result,
the Snapshot for the build was not created and we have nothing to
release. Try again.

Apparently this is a known issue but the cause is not yet known. Might
have something to do with Tekton Chains taking extremely long to attest
images (the ITS runs Enterprise Contract verification, which depends on
the existence of the attestation).

Signed-off-by: Adam Cmiel <acmiel@redhat.com>